requirements.txt
fastapi
uvicorn
Jinja2
python-dotenv
python-multipart
openai
google-cloud-firestore